classifier-service/requirements.txt
flask>=3.0,<4.0
pillow>=10.0,<12.0
torch>=2.0
torchvision>=0.15
